requirements-dev.txt
coverage
flake8
flake8-blind-except
flake8-docstrings>=1.3
mongomock
pytest
pytest-cov
tox
virtualenv
